EE-412: Move WasmTestBuilder to test support.
